Move to single-precision floats

+++ include/fizz/Constants.h
@@ -3,9 +3,9 @@
 #include <cstdint>
 
 constexpr unsigned int kTargetFPS{0};
-constexpr double kTargetFrametime{kTargetFPS ? 1.0 / static_cast<double>(kTargetFPS) : 0};
-constexpr double kPhysicStep{0.001};
+constexpr float kTargetFrametime{kTargetFPS ? 1.0 / static_cast<float>(kTargetFPS) : 0};
+constexpr float kPhysicStep{0.001};
 
-constexpr double kPi{3.141592653589793};
+constexpr float kPi{3.141592653589793};
 
 using ID = uint64_t;
